+++ src/components/StyledSearch.js
